refactor(davinci): move the S2 DOM emit driver into its own module
`emit.rs` had reached 349 of the 350-line source budget, so the next
capability could not add a field to `EmitCx` without breaking the gate.

The driver — option validation, `EmitCx` construction, preamble assembly
and the render-function body — moves verbatim to `emit/run.rs`, leaving
the module root as the capability index, the module list and the type
declarations. `emit::run` is a descendant of `emit`, so `EmitCx`'s private
fields stay reachable and no visibility widened.

No behaviour change: pure move plus the import bookkeeping it forces.
