+++ crates/vize_s1_to_s2/src/emit.rs
@@ -82,6 +82,7 @@ mod props_object;
 mod props_object_merge;
 mod props_static;
 mod props_value;
+mod run;
 mod sfc_style;
 mod slot_root;
 mod slots;
@@ -97,14 +98,13 @@ mod vnode_static;
 mod vtext;
 
 use alloc::vec::Vec as StdVec;
-use vize_davinci::diagnostic::Severity;
 use vize_davinci::id::NodeId;
 use vize_davinci::side_table::SideTable;
 use vize_s0::String;
 use vize_s2::op::Namespace;
 use vize_s2::scope::ScopeFacts;
 
-use crate::lower::{ForWrapper, Lowered, WrapperKeys};
+use crate::lower::{ForWrapper, WrapperKeys};
 use crate::pass::S2Facts;
 use crate::pass::walk::PageWalk;
 
@@ -118,9 +118,9 @@ pub use self::entry::{
     DomEmit, emit_dom_source, emit_dom_source_with_caps, emit_dom_source_with_options,
 };
 pub use self::error::{EmitError, UnsupportedReason, UnsupportedRefusal};
-use self::fragment::emit_root;
-use self::helper::Helper;
 pub use self::options::{BindingKind, BindingTable, DomEmitMode, DomEmitOptions};
+use self::run::emit_dom_with_emit_budget;
+pub use self::run::{emit_dom, emit_dom_with_options};
 
 /// Per-emit numbering + helper buffer. Page-order ids re-derive the
 /// same arithmetic the S2 passes use so compound text facts resolve.
@@ -206,132 +206,4 @@ struct EmitCx<'facts> {
     component_name: Option<&'facts str>,
     /// The transform scope and codegen slot params the prefixer consults.
     scope: prefix::PrefixScope<'facts>,
-}
-
-/// Emit a DOM render function from an already-lowered (and typically
-/// transformed) S2 artifact under the shipped default options. `facts`
-/// is the transform product compounds compile from.
-pub fn emit_dom(lowered: &Lowered<'_>, facts: &S2Facts) -> Result<DomEmit, EmitError> {
-    emit_dom_with_options(lowered, facts, &DomEmitOptions::DEFAULT)
-}
-
-/// [`emit_dom`] under explicit [`DomEmitOptions`].
-pub fn emit_dom_with_options<'f>(
-    lowered: &'f Lowered<'_>,
-    facts: &'f S2Facts,
-    options: &DomEmitOptions<'f>,
-) -> Result<DomEmit, EmitError> {
-    emit_dom_with_emit_budget(lowered, facts, options).map(|(emit, _)| emit)
-}
-
-fn emit_dom_with_emit_budget<'f>(
-    lowered: &'f Lowered<'_>,
-    facts: &'f S2Facts,
-    options: &DomEmitOptions<'f>,
-) -> Result<(DomEmit, u32), EmitError> {
-    if options.is_ts && !cfg!(feature = "typescript") {
-        return Err(EmitError::unsupported(
-            UnsupportedReason::TypeScriptLaneUnavailable,
-        ));
-    }
-    if lowered
-        .diagnostics
-        .iter()
-        .any(|diagnostic| diagnostic.severity == Severity::Error)
-    {
-        return Err(EmitError::Diagnostics);
-    }
-    // `static_cache = inline || !hoists.is_empty()`.
-    let static_cache =
-        options.inline || static_cache::enabled(&lowered.root, facts, &lowered.wrappers);
-    let mut cx = EmitCx {
-        buf: Buf::new(options.inline),
-        source: lowered.source,
-        facts,
-        scopes: &lowered.scopes,
-        wrappers: &lowered.wrappers,
-        for_wrappers: &lowered.for_wrappers,
-        walk: PageWalk::new(),
-        scope_names: StdVec::new(),
-        if_branch_key: 0,
-        once_cache_index: 0,
-        once_depth: 0,
-        once_element_depth: 0,
-        in_v_for: false,
-        conditional_v_for_item: false,
-        template_for_item_single_root: false,
-        template_for_item_root_id: None,
-        template_if_branch_root: false,
-        template_if_for_branch_root: false,
-        suppress_template_for_child_key: false,
-        skip_memo: false,
-        slot_param_depth: 0,
-        hoist_static_vnodes: false,
-        transition_slot_root: false,
-        static_cache,
-        parent_ns: Namespace::Html,
-        prefix_identifiers: options.prefix_identifiers,
-        is_ts: options.is_ts,
-        cache_handlers: options.cache_handlers,
-        hoisted_scope_id: options.hoisted_scope_id,
-        cache_sites: StdVec::new(),
-        used_unref: core::cell::Cell::new(u32::MAX),
-        component_name: options.component_name,
-        scope: prefix::PrefixScope::new(
-            options.bindings,
-            options.prefix_identifiers,
-            options.is_ts,
-            options.inline,
-        ),
-    };
-    let filters = &facts.legacy.filters;
-    if facts.legacy.filter_helper_precedes_components {
-        cx.buf.prefer(Helper::ResolveFilter);
-    }
-    let mut helper_walk = PageWalk::new();
-    let prefer_cx = helper_preference::PreferCx {
-        facts,
-        for_wrappers: &lowered.for_wrappers,
-        bindings: options.bindings,
-    };
-    helper_preference::prefer_helpers(&mut cx.buf, &prefer_cx, &mut helper_walk, &lowered.root);
-    fragment::prefer_root_fragment(&mut cx.buf, &lowered.root);
-    cx.buf
-        .push(options.mode.render_signature(options.bindings.is_some()));
-    cx.buf.indent();
-    cx.buf.newline();
-    let names = component::collect_names(&lowered.root);
-    let dirs = directive::collect_names(&lowered.root);
-    let mut resolved_assets = false;
-    if !names.is_empty() {
-        resolved_assets |= component::emit_resolves(&mut cx, &names);
-    }
-    if !dirs.is_empty() {
-        resolved_assets |= directive::emit_resolves(&mut cx, &dirs);
-    }
-    if !filters.is_empty() {
-        filter::emit_resolves(&mut cx, filters);
-        resolved_assets = true;
-    }
-    if resolved_assets {
-        cx.buf.newline();
-    }
-    cx.buf.push("return ");
-    emit_root(&mut cx, &lowered.root)?;
-    cx.buf.deindent();
-    cx.buf.newline();
-    cx.buf.push("}");
-    // `_unref` is a *transform* registration: it lists with the pre-walk's
-    // preferred helpers, at the op whose expression needed it — ahead of a
-    // structural helper that op registers later (`renderList`).
-    let unref_visit = cx.used_unref.get();
-    if unref_visit != u32::MAX {
-        cx.buf.prefer_at_visit(Helper::Unref, unref_visit);
-        cx.buf.use_helper(Helper::Unref);
-    }
-    cache_slots::renumber(&mut cx);
-    let emit_visits = cx.walk.visits();
-    let preamble = cx.buf.preamble(options);
-    let code = cx.buf.code;
-    Ok((DomEmit { preamble, code }, emit_visits))
 }
